Move subject-search to mono source set
