+++ app/shared/src/androidMain/kotlin/ui/subject/search/SubjectList.preview.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.him188.ani.app.ui.subject
+package me.him188.ani.app.ui.subject.search
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
